+++ builder_pkgs/mockito/lib/deprecated.dart
@@ -1,0 +1,5 @@
+@Deprecated('Using spy() is deprecated as we drop dart:mirrors usage')
+library mockito.deprecated;
+
+export 'mockito.dart';
+export 'src/spy.dart' show spy;