+++ builder_pkgs/mockito/lib/src/spy.dart
@@ -4,6 +4,20 @@ import 'dart:mirrors';
 
 import 'mock.dart' show CannedResponse, setDefaultResponse;
 
+@Deprecated(
+  'Mockito is dropping support for spy(). Instead it is recommended to come '
+  'up with a code-generation technique or just hand-coding a Stub object that '
+  'forwards to a delegate. For example:\n\n'
+  'class SpyAnimal implements Animal {\n'
+  '  final Animal _delegate;\n'
+  '  FakeAnimal(this._delegate);\n'
+  '\n'
+  '  @override\n'
+  '  void eat() {\n'
+  '    _delegate.eat();\n'
+  '  }\n'
+  '}'
+)
 dynamic spy(dynamic mock, dynamic spiedObject) {
   var mirror = reflect(spiedObject);
   setDefaultResponse(