+++ builder_pkgs/mockito/test/mockito_test.dart
@@ -1,7 +1,6 @@
+import 'package:mockito/mockito.dart';
+import 'package:mockito/src/mock.dart' show resetMockitoState;
 import 'package:test/test.dart';
-
-import 'package:mockito/src/mock.dart';
-import 'package:mockito/src/spy.dart';
 
 class RealClass {
   String methodWithoutArgs() => "Real";
@@ -67,24 +66,6 @@ void main() {
     // In some of the tests that expect an Error to be thrown, Mockito's
     // global state can become invalid. Reset it.
     resetMockitoState();
-  });
-
-  group("spy", () {
-    setUp(() {
-      mock = spy(new MockedClass(), new RealClass());
-    });
-
-    test("should delegate to real object by default", () {
-      expect(mock.methodWithoutArgs(), 'Real');
-    });
-    test("should record interactions delegated to real object", () {
-      mock.methodWithoutArgs();
-      verify(mock.methodWithoutArgs());
-    });
-    test("should behave as mock when expectation are set", () {
-      when(mock.methodWithoutArgs()).thenReturn('Spied');
-      expect(mock.methodWithoutArgs(), 'Spied');
-    });
   });
 
   group("mixin support", () {